Multiple entries to the index (#8)
To be able to send multiple notifications or multiple index entries in
one call, the index should be able to accept such things.

### Example:

An app wants to notify multiple users, e.g. Alice follows Bob and
Charlie in one action. She can indicate this in the graph index in one
call:

```json
[
  {
    "key": "follow",
    "value": {
      "type": "follow",
      "accountId": "bob.near"
    }
  },
  {
    "key": "follow",
    "value": {
      "type": "follow",
      "accountId": "charlie.near"
    }
  }
]
```

This will be serialized to JSON and should be added under `index/graph`:
```json
{
  "index": {
    "graph": "[{\"key\":\"follow\",\"value\":{\"type\":\"follow\",\"accountId\":\"bob.near\"}},{\"key\":\"follow\",\"value\":{\"type\":\"follow\",\"accountId\":\"charlie.near\"}}]"
  }
}
```